TG_bot/requirements.txt
anyio==3.6.2
APScheduler==3.6.3
cachetools==4.2.2
certifi==2022.12.7
emoji==2.2.0
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
mypy==0.991
mypy-extensions==0.4.3
python-telegram-bot==13.14
pytz==2022.7
pytz-deprecation-shim==0.1.0.post0
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
socksio==1.0.0
tomli==2.0.1
tornado==6.1
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
